ml-model/requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Text processing
nltk==3.8.1
textblob==0.17.1

# Web framework for API
flask==2.3.2
flask-cors==4.0.0

# Data handling
joblib==1.3.1
pickle-mixin==1.0.2

# Email processing
email-validator==2.0.0
beautifulsoup4==4.12.2
html2text==2020.1.16

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.0
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2